Conflicts:
	src/main/resources/resources/ApplicationResources_pt.properties

+++ src/main/java/net/sourceforge/fenixedu/dataTransferObject/InfoFileContent.java
@@ -46,7 +46,7 @@ public class InfoFileContent extends InfoObject {
             setMimeType(fileItem.getMimeType());
             setChecksum(fileItem.getChecksum());
             setChecksumAlgorithm(fileItem.getChecksumAlgorithm());
-            setSize(fileItem.getSize());
+            setSize(fileItem.getSize().intValue());
             setExternalStorageIdentification(fileItem.getExternalStorageIdentification());
 
         }
